+++ include/move_structure.hpp
@@ -42,7 +42,7 @@ class MoveStructure {
         // char compute_char(uint64_t idx);
 
         uint64_t LF(uint64_t row_number);
-        uint64_t LF_move(uint64_t& pointer, uint64_t& i);
+        uint16_t LF_move(uint64_t& pointer, uint64_t& i);
         uint64_t fast_forward(uint64_t& offset, uint64_t index, uint64_t x);
 
         uint64_t compute_threshold(uint64_t r_idx, uint64_t pointer, char lookup_char);
